+++ src/complete_shell.rs
@@ -75,19 +75,20 @@ where
 
         // stash old completions
         let mut comp_items = Vec::new();
-        args.swap_comps(&mut comp_items);
+        args.swap_comps_with(&mut comp_items);
 
         let res = self.inner.eval(args);
 
         // at this point comp_items contains values added by the inner parser
-        args.swap_comps(&mut comp_items);
+        args.swap_comps_with(&mut comp_items);
 
-        if let Some(comp) = &mut args.comp {
+        let depth = args.depth;
+        if let Some(comp) = args.comp_mut() {
             for ci in comp_items {
                 if let Some(is_arg) = ci.meta_type() {
-                    comp.push_shell(self.op, args.depth, is_arg);
+                    comp.push_shell(self.op, depth, is_arg);
                 } else {
-                    comp.comps.push(ci);
+                    comp.push_comp(ci);
                 }
             }
         }